refactor: adjustment of quantity reducing on batch and ad, including dynamic set quantity on creating ad

+++ src/main/java/br/com/mercadolivre/projetointegrador/marketplace/dtos/CreateOrUpdateAdDTO.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class CreateOrUpdateAdDTO {
 
   @NotEmpty(message = "O produto deve pertencer a pelo menos um lote.")
-  private List<Long> batchesId;
+  private List<Integer> batchesId;
 
   @NotEmpty(message = "O campo nome deve ser preenchido.")
   private String name;
